+++ embed_xray_other.go
@@ -1,0 +1,7 @@
+//go:build !windows
+
+package main
+
+var xrayEXE []byte
+var geoipDAT []byte
+var geositeDAT []byte